PR-15: extract gnome_ext/ (FINAL PR)
Behaviour-preserving relocation. Closes the M1 milestone.

- gnome_ext/embed.rs (cfg-gated): gnome_ext_file! macro with the
  critical relative-path bump from concat!("../../", ...) to
  concat!("../../../", ...) — one level deeper because the file
  moved one module level. All 9 EMBEDDED_* include_str! consts.
  compile_gnome_schemas + write_embedded_extension_to_dir.
- gnome_ext/detection.rs: GnomeExtensionStatus, GnomeDbusProbeResult,
  state parsers, dbus probe variants, gnome_extension_status,
  session_bus_name_has_owner (PR-08 widening's final home).
- gnome_ext/install.rs: pack/install/enable + path helpers.
- gnome_ext/mod.rs: submodule declarations + pub(crate) use re-exports
  + orchestration fns (setup_gnome_extension, ensure_gnome_extension,
  print_*).

main.rs: 958 -> 250 LOC. Pre-refactor was 7983 LOC — 96.9% reduction.
Well below the plan's <=400 LOC target.

supervisor/capabilities.rs: two call sites updated to
crate::gnome_ext::detection::session_bus_name_has_owner.

Verification: cargo build (default features) ✓; cargo build
--no-default-features ✓; cargo test --bin kanata-switcher 261/0.
M1 milestone complete.
