+++ src/daemon/gnome_ext/install.rs
@@ -1,0 +1,150 @@
+use std::env;
+use std::path::{Path, PathBuf};
+use std::process::Command;
+use crate::constants::*;
+
+pub(crate) fn get_gnome_extension_fs_path() -> PathBuf {
+    let exe_path = env::current_exe().unwrap();
+    let exe_dir = exe_path.parent().unwrap();
+    exe_dir.join("gnome")
+}
+
+pub(crate) fn gnome_extension_fs_exists() -> bool {
+    let path = get_gnome_extension_fs_path();
+    path.join("extension.js").exists()
+        && path.join("metadata.json").exists()
+        && path.join("prefs.js").exists()
+        && path.join("format.js").exists()
+        && path.join("dbus.js").exists()
+        && path.join("focus.js").exists()
+        && path.join("daemon-state.js").exists()
+        && path.join("extension-multiplex.js").exists()
+        && path.join(GNOME_EXTENSION_SCHEMA_FILE).exists()
+        && path.join(GNOME_EXTENSION_SCHEMA_COMPILED).exists()
+}
+
+pub(crate) fn pack_and_install_from_dir(src_dir: &Path, tmp_dir: &Path) -> Result<(), String> {
+    let zip_name = format!("{}.shell-extension.zip", GNOME_EXTENSION_UUID);
+
+    let pack_result = Command::new("gnome-extensions")
+        .args([
+            "pack",
+            src_dir.to_str().unwrap(),
+            "--force",
+            &format!("--out-dir={}", tmp_dir.display()),
+        ])
+        .output();
+
+    if pack_result.is_err() || !pack_result.as_ref().unwrap().status.success() {
+        return Err("gnome-extensions pack failed".to_string());
+    }
+
+    let zip_path = tmp_dir.join(&zip_name);
+    let install_result = Command::new("gnome-extensions")
+        .args(["install", zip_path.to_str().unwrap(), "--force"])
+        .output();
+
+    if install_result.is_err() || !install_result.as_ref().unwrap().status.success() {
+        return Err("gnome-extensions install failed".to_string());
+    }
+
+    Ok(())
+}
+
+#[allow(unused_variables, unused_assignments)]
+pub(crate) fn install_gnome_extension() -> bool {
+    use std::fs;
+    let tmp_dir = tempfile::tempdir().unwrap();
+    let fs_path = get_gnome_extension_fs_path();
+    let mut fs_error: Option<String> = None;
+
+    // Try filesystem first
+    if gnome_extension_fs_exists() {
+        println!("[GNOME] Installing from filesystem: {}", fs_path.display());
+        match pack_and_install_from_dir(&fs_path, tmp_dir.path()) {
+            Ok(()) => {
+                println!("[GNOME] Extension installed");
+                return true;
+            }
+            Err(e) => {
+                eprintln!("[GNOME] Failed to install from filesystem: {}", e);
+                fs_error = Some(e);
+            }
+        }
+    } else {
+        eprintln!(
+            "[GNOME] Extension files not found at filesystem path: {}",
+            fs_path.display()
+        );
+    }
+
+    // Fallback to embedded extension
+    #[cfg(feature = "embed-gnome-extension")]
+    {
+        use super::embed::write_embedded_extension_to_dir;
+        eprintln!("[GNOME] Falling back to embedded extension...");
+        let embedded_dir = tmp_dir.path().join("embedded");
+        fs::create_dir_all(&embedded_dir).unwrap();
+
+        if let Err(e) = write_embedded_extension_to_dir(&embedded_dir) {
+            eprintln!("[GNOME] Failed to write embedded extension: {}", e);
+            super::print_gnome_extension_install_instructions(
+                "Auto-install failed: could not write embedded extension files.",
+            );
+            return false;
+        }
+
+        match pack_and_install_from_dir(&embedded_dir, tmp_dir.path()) {
+            Ok(()) => {
+                println!("[GNOME] Extension installed (from embedded)");
+                return true;
+            }
+            Err(e) => {
+                eprintln!("[GNOME] Failed to install from embedded: {}", e);
+                super::print_gnome_extension_install_instructions(&format!("Auto-install failed: {}", e));
+                return false;
+            }
+        }
+    }
+
+    #[cfg(not(feature = "embed-gnome-extension"))]
+    {
+        let reason = if let Some(e) = fs_error {
+            format!(
+                "Found extension files at {}, but installation failed: {}. \
+                 Cannot fall back to embedded extension (disabled in this build).",
+                fs_path.display(),
+                e
+            )
+        } else {
+            "Extension files not found and embedded extension is disabled in this build."
+                .to_string()
+        };
+        super::print_gnome_extension_install_instructions(&reason);
+        return false;
+    }
+}
+
+pub(crate) fn enable_gnome_extension() -> bool {
+    let result = Command::new("gnome-extensions")
+        .args(["enable", GNOME_EXTENSION_UUID])
+        .output();
+
+    match result {
+        Ok(output) if output.status.success() => {
+            println!("[GNOME] Extension enabled");
+            true
+        }
+        _ => {
+            eprintln!("[GNOME] Failed to enable extension");
+            eprintln!("[GNOME] Try restarting GNOME Shell first:");
+            eprintln!("[GNOME]   - Press Alt+F2, type \"r\", press Enter (X11 only)");
+            eprintln!("[GNOME]   - Or log out and log back in (Wayland)");
+            eprintln!(
+                "[GNOME] Then run: gnome-extensions enable {}",
+                GNOME_EXTENSION_UUID
+            );
+            false
+        }
+    }
+}
