+++ src/daemon/supervisor/capabilities.rs
@@ -6,8 +6,8 @@ use crate::errors::DynError;
 pub(crate) async fn detect_desktop_capabilities() -> Result<DesktopCapabilities, DynError> {
     let connection = Connection::session().await?;
     let dbus = zbus::fdo::DBusProxy::new(&connection).await?;
-    let gnome_owner = crate::session_bus_name_has_owner(&dbus, GNOME_SHELL_BUS_NAME).await;
-    let kde_owner = crate::session_bus_name_has_owner(&dbus, KDE_KWIN_BUS_NAME).await;
+    let gnome_owner = crate::gnome_ext::detection::session_bus_name_has_owner(&dbus, GNOME_SHELL_BUS_NAME).await;
+    let kde_owner = crate::gnome_ext::detection::session_bus_name_has_owner(&dbus, KDE_KWIN_BUS_NAME).await;
     Ok(DesktopCapabilities {
         gnome_owner,
         kde_owner,
